Update to work with the last rust syntax and nphysics api.
This also removes the position correction from the CPU version, which is fair because there is no
such correction on the GPU version.

All those changes make the CPU version a *lot* faster (almost 10 times faster!).
GPU version is still a bit faster but the gap seems to be very small.